Refactored retry on its own file
